Yolov7_pose_test/requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
backports.zoneinfo==0.2.1
bcrypt==4.1.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
easyocr==1.7.1
exceptiongroup==1.2.0
extra-streamlit-components==0.1.60
fastapi==0.104.1
filelock==3.13.1
fonttools==4.45.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httptools==0.6.1
idna==3.6
imageio==2.33.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.4
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
protobuf==4.25.1
pyarrow==14.0.1
pycairo==1.24.0
pyclipper==1.3.0.post5
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
PyGObject==3.44.1
PyJWT==2.8.0
pyparsing==3.1.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.31.1
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.13.0
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.27.0
streamlit==1.28.2
streamlit-aggrid==0.3.4.post3
streamlit-authenticator==0.2.3
streamlit-cropper==0.2.2
streamlit-option-menu==0.3.6
streamlit-toggle-switch==1.0.2
sympy==1.12
tenacity==8.2.3
tifffile==2023.7.10
toml==0.10.2
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tornado==6.4
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
websockets==12.0
zipp==3.17.0
